Refactored the MetadataIndex to extend UnisolatedBTree and put it to some use in the MetadataService and validated aspects of its use in the test case for the metadata service.
git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/trunk@203 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata/src/java/com/bigdata/scaleup/ResourceState.java
@@ -65,9 +65,9 @@ public enum ResourceState {
     
     public String toString() {return name;}
     
-    public int valueOf() {return id;}
+    public short valueOf() {return (short)id;}
     
-    static public ResourceState valueOf(int id) {
+    static public ResourceState valueOf(short id) {
         switch(id) {
         case 0: return New;
         case 1: return Live;
